remise en forme du code
