+++ src/feats/courses/functions.ts
@@ -4,9 +4,11 @@ import { getOneAsynchronous } from "../asynchronous-courses/function";
 import { lessons } from "../asynchronous-courses/schema";
 import { classes, synchronousCourses } from "../synchronous-courses/schema";
 import { courses } from "./schema";
-import { serverOnly$ } from "vite-env-only";
 
-export type CourseCardAsync =  Omit<InferSelectModel<typeof lessons>, "id" | "creationTime" | "updateTime" | "title" | "description"> & {
+export type CourseCardAsync = Omit<
+	InferSelectModel<typeof lessons>,
+	"id" | "creationTime" | "updateTime" | "title" | "description"
+> & {
 	chapterId: number;
 	capsuleUrl: string;
 	previewable: boolean;
@@ -31,8 +33,7 @@ export type Course = {
 };
 
 export type CourseCard = Omit<
-	InferSelectModel<typeof classes> &
-		InferSelectModel<typeof lessons>,
+	InferSelectModel<typeof classes> & InferSelectModel<typeof lessons>,
 	"id" | "creationTime" | "updateTime"
 > &
 	Course &
@@ -41,4 +42,25 @@ export type CourseCard = Omit<
 export async function getAllCours() {
 	const result = await db.select().from(courses);
 	return result;
+}
+
+export async function getOneCours(id: number) {
+	const course = await db.select().from(courses).where(eq(courses.id, id));
+
+	const syncCourse: CourseCardSync[] = await db
+		.select()
+		.from(classes)
+		.where(eq(synchronousCourses.id, id));
+
+	let CoursInfo;
+
+	if (syncCourse.length === 0) {
+		const asyncCourseInfo: CourseCardAsync[] = await getOneAsynchronous(id);
+
+		CoursInfo = [...course, ...asyncCourseInfo];
+	} else {
+		CoursInfo = [...course, ...syncCourse];
+	}
+
+	return CoursInfo;
 }
